feat!: Migrate module `utils.template` to `slack_messages.message_templates`
