--- django_slack_tools/utils/template/__init__.py
@@ -1,5 +1,0 @@
-from .base import BaseTemplate
-from .dict import DictTemplate
-from .django import DjangoTemplate
-
-__all__ = ("BaseTemplate", "DictTemplate", "DjangoTemplate")
